+++ docker/publishcontainer.sh
@@ -2,7 +2,7 @@
 
 ver=`../blobfuse2 --version | cut -d " " -f 3`
 image="azure-blobfuse2.$ver"
-alpine_image="azure-alpine-blobfuse2.$ver"
+mariner_image="azure-mariner-blobfuse2.$ver"
 
 docker login blobfuse2containers.azurecr.io --username $1 --password $2
 
@@ -10,9 +10,9 @@ docker login blobfuse2containers.azurecr.io --username $1 --password $2
 docker tag $image:latest blobfuse2containers.azurecr.io/$image
 docker push blobfuse2containers.azurecr.io/$image
 
-# Publish Alpine container image
-docker tag $alpine_image:latest blobfuse2containers.azurecr.io/$alpine_image
-docker push blobfuse2containers.azurecr.io/$alpine_image
+# Publish Mariner container image
+docker tag $mariner_image:latest blobfuse2containers.azurecr.io/$mariner_image
+docker push blobfuse2containers.azurecr.io/$mariner_image
 
 docker logout blobfuse2containers.azurecr.io